Update package to version 0.1.4 and expand HugoToolkit functionality
This update includes increasing the version of the package from 0.1.3 to 0.1.4. Extensive new methods and functionality have been added to the HugoToolkit class, mainly focused on generating various types of content, such as mini portfolios and galleries from Wasabi, YouTube embeds, and more. Additionally, this commit involves creating a test suite for the new HugoToolkit functionality, along with some minor changes to improve the clarity of a function docstring in "file_utils.py".